requirements.txt
# Core ML and Data Science
scikit-learn==1.5.2
pandas>=1.5.0
numpy>=1.21.0
joblib>=1.2.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Environment and Configuration
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Task Queue and Background Jobs
celery>=5.3.0
redis>=4.5.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
aiosqlite>=0.19.0

# Async Support
anyio>=3.7.0
asyncio-mqtt>=0.16.0

# File Handling
aiofiles>=23.0.0

# Performance and Monitoring
slowapi>=0.1.0
prometheus-client>=0.17.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Logging and Error Handling
structlog>=23.0.0
sentry-sdk>=1.40.0

# Security
cryptography>=41.0.0
passlib>=1.7.0
python-jose>=3.3.0

# HTTP Client
httpx>=0.24.0
aiohttp>=3.8.0

# Serialization
pydantic>=2.0.0
marshmallow>=3.19.0

# Utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0